code cleanup and seperated repsonse types to their own file
