+++ ipfs/response_structs.go
@@ -1,0 +1,14 @@
+package ipfs
+
+type DownloadResponse struct {
+	Status    int
+	FileData  []byte
+	NameSpace string
+	Path      string
+}
+
+type UploadResponse struct {
+	Status int
+	Cid    string
+	Path   string
+}
